gpu: nvgpu: fix MISRA issues nvgpu.hal.fifo.pbdma
MISRA Rule 10.3 doesn't allow implicit conversion between essential
types.
This patch typecasts return value of pbdma_syncpointb_syncpt_index_v().

Jira NVGPU-3767

Reviewed-on: https://git-master.nvidia.com/r/2147802
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/hal/fifo/pbdma_gm20b.c
@@ -136,7 +136,7 @@ void gm20b_pbdma_syncpoint_debug_dump(struct gk20a *g,
 			info->hw_state.pending_acquire ? "Waiting" : "Waited",
 			pbdma_syncpointb_syncpt_index_v(syncpointb),
 			nvgpu_nvhost_syncpt_get_name(g->nvhost_dev,
-				pbdma_syncpointb_syncpt_index_v(syncpointb)),
+			(int) pbdma_syncpointb_syncpt_index_v(syncpointb)),
 			pbdma_syncpointa_payload_v(syncpointa));
 	}
 #endif
